requirements.txt
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=14.0.0
pandas-ta>=0.3.14b
scipy>=1.11.0
groq>=0.11.0
streamlit>=1.35.0
plotly>=5.20.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
tenacity>=8.2.0
